user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
]
